Extract package file finding to internal/packages
